Move `RoomFlowCoordinator` state machine elements to a separate file.
